lead_manager/requirements.txt
flask>=3.0
google-api-python-client>=2.0
google-auth>=2.0
python-dotenv>=1.0.0
gunicorn>=21.2
requests>=2.31.0
